Moved the create team form.
